change in the delete

+++ backend/controller/product.controller.js
@@ -35,6 +35,13 @@ export const createProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
 
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(404).json({
+      success: false,
+      message: "Invalid product ID",
+    });
+  }
+
   try {
     await Product.findByIdAndDelete(id);
     res.status(200).json({
@@ -42,9 +49,9 @@ export const deleteProduct = async (req, res) => {
       message: "successfully deleted",
     });
   } catch (error) {
-    res.status(404).json({
+    res.status(500).json({
       success: false,
-      message: "product not found",
+      message: "server error",
     });
   }
 };
